Fix replacing files that rails new uses as templates
You can't just use `copy_file Gemfile`, as that will copy in the default
source that `rails new` uses as a template, which includes stuff to do
with which gems should be included based on the platform.

We want to copy in our own Gemfile, so it has to be at a different
relative path than the default Rails one, so I've moved it to
'overrides/' instead so we can refer to it correctly.